Switch from MD2 to MD5 models.

+++ src/WorldClient.cpp
@@ -81,7 +81,9 @@ void CWorldClient::AddWorldToHistory()
     clstate.localtime = CLynxSys::GetTicks();
     m_history.push_front(clstate);
     if(m_history.size() > 80)
-        fprintf(stderr, "Client snapshot history buffer size is getting larger: %i\n", m_history.size());
+        fprintf(stderr,
+                "Client snapshot history buffer size is getting larger: %u\n",
+                (uint32_t)m_history.size());
     while(clstate.localtime - m_history.back().localtime > MAX_CLIENT_HISTORY)
         m_history.pop_back();
     while(m_history.size() > 80)
